+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
+- Add `env`, which is a internal sub-package.
 - Move `timestamp` into the `internal` directory.
 - Add `timestamp` sub-package (#2).
 - Add the build status badge to `README.md`